Fix build with ImageMagick >= 6.6.2 (closes #409).

+++ bitmap.c
@@ -187,7 +187,7 @@ bool cText2SkinBitmap::LoadNonXpm(const char *Filename, int height, int width, i
 				return false;
 			}
 			*/
-			bmp = new cBitmap(w, h, std::min((*it).depth(), 8U));
+			bmp = new cBitmap(w, h, std::min(size_t((*it).depth()), size_t(8)));
 			//Dprintf("this image has %d colors\n", (*it).totalColors());
 
 			const PixelPacket *pix = (*it).getConstPixels(0, 0, w, h);
